#1186 Heikin-Ashi bar aggregator
